Fix stack allocation
Refactor stack allocation to work correctly with locals and procedure
calls, and simplify the Frame API with calls to offset manipulations.

Fix constant popagation for logical operations NOT, AND, OR.

Fix halting of recursion for unary ops as with binary ones.

Add more unit tests for unary and binary ops